+++ Sources/Sentry/SentryANRTracker.m
@@ -6,50 +6,47 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-#if SENTRY_HAS_UIKIT
-
 @interface
 SentryANRTracker ()
 
-@property (weak, nonatomic) id<SentryANRTrackerDelegate> delegate;
-@property (nonatomic, assign) NSTimeInterval timeoutInterval;
 @property (nonatomic, strong) id<SentryCurrentDateProvider> currentDate;
 @property (nonatomic, strong) SentryCrashWrapper *crashWrapper;
 @property (nonatomic, strong) SentryDispatchQueueWrapper *dispatchQueueWrapper;
 @property (nonatomic, strong) SentryThreadWrapper *threadWrapper;
+@property (nonatomic, strong) NSMutableSet<id<SentryANRTrackerDelegate>> *listeners;
+@property (nonatomic, assign) NSTimeInterval timeoutInterval;
 
 @property (weak, nonatomic) NSThread *thread;
 
 @end
 
-@implementation SentryANRTracker
+@implementation SentryANRTracker {
+    NSObject *threadLock;
+    BOOL running;
+}
 
-- (instancetype)initWithDelegate:(id<SentryANRTrackerDelegate>)delegate
-           timeoutIntervalMillis:(NSUInteger)timeoutIntervalMillis
-             currentDateProvider:(id<SentryCurrentDateProvider>)currentDateProvider
-                    crashWrapper:(SentryCrashWrapper *)crashWrapper
-            dispatchQueueWrapper:(SentryDispatchQueueWrapper *)dispatchQueueWrapper
-                   threadWrapper:(SentryThreadWrapper *)threadWrapper
+- (instancetype)initWithTimeoutInterval:(NSTimeInterval)timeoutInterval
+                    currentDateProvider:(id<SentryCurrentDateProvider>)currentDateProvider
+                           crashWrapper:(SentryCrashWrapper *)crashWrapper
+                   dispatchQueueWrapper:(SentryDispatchQueueWrapper *)dispatchQueueWrapper
+                          threadWrapper:(SentryThreadWrapper *)threadWrapper
 {
     if (self = [super init]) {
-        self.delegate = delegate;
-        self.timeoutInterval = (double)timeoutIntervalMillis / 1000;
+        self.timeoutInterval = timeoutInterval;
         self.currentDate = currentDateProvider;
         self.crashWrapper = crashWrapper;
         self.dispatchQueueWrapper = dispatchQueueWrapper;
         self.threadWrapper = threadWrapper;
+        self.listeners = [NSMutableSet new];
+        threadLock = [[NSObject alloc] init];
+        running = false;
     }
     return self;
 }
 
-- (void)start
-{
-    [NSThread detachNewThreadSelector:@selector(detectANRs) toTarget:self withObject:nil];
-}
-
 - (void)detectANRs
 {
-    NSThread.currentThread.name = @"io.sentry.anr-tracker";
+    NSThread.currentThread.name = @"io.sentry.app-hang-tracker";
 
     self.thread = NSThread.currentThread;
 
@@ -68,7 +65,7 @@ SentryANRTracker ()
         if (blockExecutedOnMainThread) {
             if (wasPreviousANR) {
                 [SentryLog logWithMessage:@"ANR stopped." andLevel:kSentryLevelWarning];
-                [self.delegate anrStopped];
+                [self ANRStopped];
             }
 
             wasPreviousANR = NO;
@@ -103,18 +100,85 @@ SentryANRTracker ()
 
         wasPreviousANR = YES;
         [SentryLog logWithMessage:@"ANR detected." andLevel:kSentryLevelWarning];
-        [self.delegate anrDetected];
+        [self ANRDetected];
+    }
+}
+
+- (void)ANRDetected
+{
+    NSArray *localListeners;
+    @synchronized(self.listeners) {
+        localListeners = [self.listeners allObjects];
+    }
+
+    for (id<SentryANRTrackerDelegate> target in localListeners) {
+        [target anrDetected];
+    }
+}
+
+- (void)ANRStopped
+{
+    NSArray *targets;
+    @synchronized(self.listeners) {
+        targets = [self.listeners allObjects];
+    }
+
+    for (id<SentryANRTrackerDelegate> target in targets) {
+        [target anrStopped];
+    }
+}
+
+- (void)addListener:(id<SentryANRTrackerDelegate>)listener
+{
+    @synchronized(self.listeners) {
+        [self.listeners addObject:listener];
+
+        if (self.listeners.count > 0 && !running) {
+            @synchronized(threadLock) {
+                if (!running) {
+                    [self start];
+                }
+            }
+        }
+    }
+}
+
+- (void)removeListener:(id<SentryANRTrackerDelegate>)listener
+{
+    @synchronized(self.listeners) {
+        [self.listeners removeObject:listener];
+
+        if (self.listeners.count == 0) {
+            [self stop];
+        }
+    }
+}
+
+- (void)clear
+{
+    @synchronized(self.listeners) {
+        [self.listeners removeAllObjects];
+        [self stop];
+    }
+}
+
+- (void)start
+{
+    @synchronized(threadLock) {
+        [NSThread detachNewThreadSelector:@selector(detectANRs) toTarget:self withObject:nil];
+        running = YES;
     }
 }
 
 - (void)stop
 {
-    [SentryLog logWithMessage:@"Stopping ANR detection" andLevel:kSentryLevelInfo];
-    [self.thread cancel];
+    @synchronized(threadLock) {
+        [SentryLog logWithMessage:@"Stopping ANR detection" andLevel:kSentryLevelInfo];
+        [self.thread cancel];
+        running = NO;
+    }
 }
 
 @end
-
-#endif
 
 NS_ASSUME_NONNULL_END