+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -8,6 +8,7 @@
 #import "NSURLProtocolSwizzle.h"
 #import "PrivateSentrySDKOnly.h"
 #import "SentryANRTracker.h"
+#import "SentryANRTrackingIntegration.h"
 #import "SentryAppStartMeasurement.h"
 #import "SentryAppStartTracker.h"
 #import "SentryAppStartTrackingIntegration.h"
@@ -162,7 +163,6 @@
 #import "URLSessionTaskMock.h"
 
 #if SENTRY_HAS_UIKIT
-#    import "SentryANRTrackingIntegration.h"
 #    import "SentryUIEventTracker.h"
 #    import "SentryUIEventTrackingIntegration.h"
 #endif